exp: refactor stash handling into separate module
